+++ Synthesizer/Oscillators/Base Implementations/SineOscillator.h
@@ -5,10 +5,10 @@
 class SineOscillator : public Oscillator {
 public:
     SineOscillator();
-    SineOscillator(double hz, double sample_rate);
+    SineOscillator(float hz, float sample_rate);
     SineOscillator(const SineOscillator& other);
 
-    double get_sample() override;
+    float get_sample() override;
 };
